Add documentation link for compile pipeline in upload pipeline dialog
